requirements.txt
# Core MCP and FastAPI dependencies
fastmcp>=2.12.0
httpx>=0.27
python-dotenv>=1.0
google-auth>=2.31
openai>=1.40.0
fastapi>=0.95.0
langfuse>=0.10.0
pydantic>=2.4.0
uvicorn[standard]

# Discord Bot
discord.py>=2.3.0

# Microsoft Office
python-docx>=0.8.11

# Social Media APIs
TikTokApi>=6.0.0
praw>=7.7.0
telethon>=1.28.0

# Payment Processing
stripe>=7.0.0

# Google Services
googlemaps>=4.10.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0

# Web Scraping & Browser Automation
beautifulsoup4>=4.12.0
hyperbrowser>=0.55.0

# Additional API Libraries
requests>=2.31.0
aiohttp>=3.8.0